commit con prodotto+req+dto+repositories abbastanza finiti

+++ src/main/java/com/betacom/backend/repositories/IProdottoRepositories.java
@@ -3,5 +3,5 @@ package com.betacom.backend.repositories;
 import com.betacom.backend.model.Prodotto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProdottoRepositories extends JpaRepository<Prodotto,Integer> {
+public interface IProdottoRepositories extends JpaRepository<Prodotto,Long> {
 }
